empty rooms now pad the channel when anyone joins an empty channel, instead of on lock
