+++ src/listeners/userListener.ts
@@ -3,9 +3,16 @@ import {
   isLockedVoiceChannel,
   unlockChannel,
   isPracticeChannel,
+  getNewUnlockedChannelName,
 } from '../utils/discordUtils/channels';
 import { isHost } from '../utils/discordUtils/users';
 import { cleanChannels } from '../utils/discordUtils/misc';
+import { MAX_EMPTY_UNLOCKED_ROOMS } from '../utils/discordUtils/constants';
+import {
+  getPracticeCategoryVoiceChannels,
+  getPracticeCategory,
+} from '../utils/discordUtils/categories';
+import { environment } from '../environment';
 
 export function listenForUsers(discord: Discord.Client) {
   // eslint-disable-next-line sonarjs/cognitive-complexity
@@ -14,9 +21,11 @@ export function listenForUsers(discord: Discord.Client) {
       return;
     }
 
+    const manager = newMember.guild.channels;
+
     // Unlock if host leaves
     if (prevMember.channel && prevMember.member && isHost(prevMember.member, prevMember.channel)) {
-      await unlockChannel(newMember.guild.channels, prevMember.channel);
+      await unlockChannel(manager, prevMember.channel);
     }
 
     // Unmute if in unlocked server
@@ -36,7 +45,33 @@ export function listenForUsers(discord: Discord.Client) {
 
     // Clean up empty voice channels
     if (prevMember.channel && isPracticeChannel(prevMember.channel)) {
-      await cleanChannels(newMember.guild.channels);
+      await cleanChannels(manager);
+    }
+
+    // Pad category with empty rooms
+    const existingVoiceChannels = getPracticeCategoryVoiceChannels(manager);
+    if (!existingVoiceChannels) {
+      return;
+    }
+    const unlockedChannels = existingVoiceChannels?.filter((vc) => !isLockedVoiceChannel(vc));
+    const emptyUnlockedChannels = unlockedChannels?.filter((vc) => vc.members.size === 0);
+
+    if (emptyUnlockedChannels.size < MAX_EMPTY_UNLOCKED_ROOMS) {
+      const practiceCategory = getPracticeCategory(manager);
+      const newChannelName = getNewUnlockedChannelName(unlockedChannels.map((c) => c.name));
+      if (!newChannelName) {
+        return;
+      }
+      await manager.create(newChannelName, {
+        type: 'voice',
+        parent: practiceCategory,
+        bitrate: environment.default_bitrate * 1000,
+      });
+      await manager.create(newChannelName, {
+        type: 'text',
+        parent: practiceCategory,
+        bitrate: environment.default_bitrate * 1000,
+      });
     }
   });
 }
