+++ src/utils/discordUtils/constants.ts
@@ -1,2 +1,2 @@
 export const VC_IDENTIFERS = ['🎸', '🎹', '🎤', '🎼', '🥁', '🍆', '🍑'];
-export const MAX_EMPTY_UNLOCKED_ROOMS = 3;
+export const MAX_EMPTY_UNLOCKED_ROOMS = 5;
